consolidate useForza hook
consolidate hook into context file
