+++ ForzaUtils/src/context/viewModels/GripViewModel.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo } from "react";
-import { useForzaData } from "../../hooks/useForzaData";
+import { useForzaData } from "../../context/Forza";
 import { TireData } from "ForzaTelemetryApi";
