--- ForzaUtils/src/hooks/useForzaData.tsx
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { ForzaContext, ForzaData } from "../context/Forza";
-
-export function useForzaData(): ForzaData {
-  return useContext(ForzaContext);
-}
